+++ .idea/libraries/Google_App_Engine_SDK.xml
@@ -1,0 +1,7 @@
+<component name="libraryTable">
+  <library name="Google App Engine SDK" type="python">
+    <CLASSES />
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
