+++ database/Base.py
@@ -1,0 +1,17 @@
+from sqlalchemy import create_engine
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import scoped_session, sessionmaker
+
+from config import DATABASE_URI
+
+engine = create_engine(DATABASE_URI)
+db_session = scoped_session(sessionmaker(autocommit=False,
+                                         autoflush=False,
+                                         bind=engine))
+Base = declarative_base()
+
+
+def init_db():
+
+    Base.query = db_session.query_property()
+    Base.metadata.create_all(bind=engine)
