+++ database/User.py
@@ -1,0 +1,15 @@
+from sqlalchemy import Column, Integer, String
+
+from database.Base import Base
+
+
+class User(Base):
+    __tablename__ = 'users'
+
+    id = Column(Integer, primary_key=True)
+    github_access_token = Column(String(255))
+    github_id = Column(Integer)
+    github_login = Column(String(255))
+
+    def __init__(self, github_access_token):
+        self.github_access_token = github_access_token
